chore: Change version to 0.1.0
After discussions with the nf-core maintainers:

Essentially that we're not dissuading moving pipelines to nf-core or
cutting releases, but major releases should meet the majority of nf-core
guidelines. Not the promise of we'll nf-corize. But cutting a 0. release
allows users to start using the pipeline and not have a HUGE single
changelog. This also allows easy reproduction for users at a later time.